refactor: some more
